refactor(ValidateAmount) : 패키지 이동
