+++ src/main/java/lotto/domain/validation/amount/ValidateAmountZeroMinus.java
@@ -1,4 +1,4 @@
-package lotto.domain.validation.AmountValidation;
+package lotto.domain.validation.amount;
 
 import lotto.exception.InputException;
